Finish up create disks.
The API only supports a single disc at a time so I've ensured it's only
a single disc at a time. I've also removed all the []*Disc and changed
it to []Disc because slices of pointers are bad or something.

This has introduced some code smell I'm too wiped to deal with right
now.

Going to bed.