Ensure prompting & --force on £££/scary commands
All destructive commands (according to a list in quality_test.go) are
now tested to ensure they have a --force flag. In addition to this I've
manually ensured that all such commands abort if --force isn't set and
the user doesn't respond with a yes.

This commit fixes the following on our private instance: #15 and #18