Fix error when bmbilling acc has no bigv acc.
Output a nice error when it is not possible to determine a default
account "name" (bigv account name), due to the default bmbilling account
not having an associated bigv account. This replaces the previous
behaviour, which bizarrely threw away the endpoint and attempted to
connect to localhost.

Also, ensure that talking to billing is shortcutted when using the int
endpoint, and hard-code the default account to bytemark when on int.