Prevent account secrets being written to debug.log
This is a bit of a sledgehammer - it temporarily sets the log.LogFile to
nil to prevent output while the HTTP requests for CreateCreditCard and
RegisterNewAccount are running.

In the future I would like to rid myself of util/log, and replace it
with the built-in log package and an io.Writer that would replace
"card_reference":"[^"]*" with "card_reference":"redacted" and such.

Closes #60 and closes #59