Add .gometalinter.json, disable staticcheck ST1005
ST1005 is the code for "error strings should not be capitalized".
error strings in bytemark-client are frequently intended for eventual
consumption by the customer, so capitalising them is ok.

A better solution in future would be to add a PrettyError type that has
a PrettyError method for this purpose (but also implements error) - but
that ship has somewhat sailed with bytemark-client, at least for now.